status/functions/pmc_status_delivery/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
azure-storage-queue
repoaudit